refactoring to lib/nl with utils and constants
