+++ server/lib/nl/nl_data_spec.py
@@ -18,8 +18,8 @@ from typing import Dict, List
 import logging
 import pandas as pd
 
-from lib.nl_detection import ClassificationType, Detection
-from lib import nl_variable, nl_topic
+from lib.nl.nl_detection import ClassificationType, Detection
+from lib.nl import nl_variable, nl_topic
 import services.datacommons as dc
