+++ server/lib/nl/nl_topic.py
