+++ server/lib/nl/nl_variable.py
