+++ server/routes/nl_interface.py
@@ -20,16 +20,16 @@ import json
 import flask
 from flask import Blueprint, current_app, render_template, escape, request
 from google.protobuf.json_format import MessageToJson, ParseDict
-from lib.nl_detection import ClassificationType, ContainedInPlaceType, Detection, NLClassifier, Place, PlaceDetection, SVDetection, SimpleClassificationAttributes
+from lib.nl.nl_detection import ClassificationType, ContainedInPlaceType, Detection, NLClassifier, Place, PlaceDetection, SVDetection, SimpleClassificationAttributes
 from typing import Dict, Union
 import pandas as pd
 import re
 import requests
 
 import services.datacommons as dc
-import lib.nl_data_spec as nl_data_spec
-import lib.nl_page_config as nl_page_config
-import lib.nl_variable as nl_variable
+import lib.nl.nl_data_spec as nl_data_spec
+import lib.nl.nl_page_config as nl_page_config
+import lib.nl.nl_variable as nl_variable
 from config import subject_page_pb2
 
 bp = Blueprint('nl', __name__, url_prefix='/nl')
